Refactor and moved auction models to auction folder
Refactor and moved auction models to auction folder
