+++ app/shared/com/ortb/demadSidePlatforms/DemandSidePlatformStaticBidResponseLogicImplementation.scala
@@ -4,7 +4,8 @@ import shared.com.ortb.demadSidePlatforms.traits.logics.DemandSidePlatformStatic
 import shared.com.ortb.demadSidePlatforms.traits.properties.{ DemandSidePlatformBiddingProperties, DemandSidePlatformCorePropertiesContracts }
 import shared.com.ortb.model.auctionbid.biddingRequests.ImpressionModel
 import shared.com.ortb.model
-import shared.com.ortb.model._
+import shared.com.ortb.model.{ auctionbid, _ }
+import shared.com.ortb.model.auctionbid.{ DemandSidePlatformBidResponseModel, ImpressionDealModel }
 import shared.com.ortb.model.auctionbid.bidresponses.BidResponseModelWrapper
 import shared.com.ortb.model.config.DemandSidePlatformConfigurationModel
 import shared.com.ortb.model.results.DemandSidePlatformBiddingRequestModel
@@ -28,7 +29,7 @@ class DemandSidePlatformStaticBidResponseLogicImplementation
     for (impression <- impressions) {
       if (impression.bidfloor.isDefined) {
         val deal : Double = impression.bidfloor.get + defaultIncrementNumber
-        val impressionDealModel = model.ImpressionDealModel(impression, deal)
+        val impressionDealModel = ImpressionDealModel(impression, deal)
         deals.addOne(impressionDealModel)
 
         val callStackModel = CallStackModel(
@@ -40,7 +41,7 @@ class DemandSidePlatformStaticBidResponseLogicImplementation
 
         callStacks.addOne(callStackModel)
       } else {
-        deals.addOne(model.ImpressionDealModel(impression, defaultStaticDeal))
+        deals.addOne(auctionbid.ImpressionDealModel(impression, defaultStaticDeal))
 
         val callStackModel = CallStackModel(
           deal = defaultStaticDeal,
@@ -54,7 +55,7 @@ class DemandSidePlatformStaticBidResponseLogicImplementation
     }
 
     val dspBidderResultModel =
-      model.DemandSidePlatformBidResponseModel(
+      DemandSidePlatformBidResponseModel(
         request,
         request.bidRequestModel,
         bidResponseWrapper = null,
