+++ app/shared/com/ortb/demadSidePlatforms/ImpressionDealsGetter.scala
@@ -1,6 +1,7 @@
 package shared.com.ortb.demadSidePlatforms
 
 import shared.com.ortb.model._
+import shared.com.ortb.model.auctionbid.{ BidFailedInfoModel, ImpressionBiddableInfoModel, ImpressionDealModel }
 import shared.com.ortb.model.config.RangeModel
 import shared.com.ortb.model.results.DemandSidePlatformBiddingRequestModel
 import shared.io.helpers.EmptyValidateHelper
