+++ app/shared/com/ortb/demadSidePlatforms/traits/logics/DemandSidePlatformStaticBidResponseLogic.scala
@@ -1,6 +1,6 @@
 package shared.com.ortb.demadSidePlatforms.traits.logics
 
-import shared.com.ortb.model.DemandSidePlatformBidResponseModel
+import shared.com.ortb.model.auctionbid.DemandSidePlatformBidResponseModel
 import shared.com.ortb.model.results.DemandSidePlatformBiddingRequestModel
 
 trait DemandSidePlatformStaticBidResponseLogic {
