+++ app/shared/com/ortb/model/EntityWrapperWithJoinedTableRowsModel.scala
@@ -3,7 +3,7 @@ package shared.com.ortb.model
 import shared.com.ortb.model.wrappers.persistent.EntityWrapper
 import shared.io.helpers.EmptyValidateHelper
 
-case class EntityWrapperWithJoinedTableRows[TBase, TBaseKey, TChildRowsType, TChildKey](
+case class EntityWrapperWithJoinedTableRowsModel[TBase, TBaseKey, TChildRowsType, TChildKey](
   row : Option[EntityWrapper[TBase, TBaseKey]],
   innerRows : Option[Seq[EntityWrapper[TChildRowsType, TChildKey]]]
 ) {
